WIP: Extract common methods to BaseAddActivity
