+++ app/src/main/java/com/arkadiusz/dayscounter/ui/addeditevent/AddActivity.kt
@@ -2,120 +2,42 @@ package com.arkadiusz.dayscounter.ui.addeditevent
 
 import android.Manifest
 import android.app.Activity
-import android.app.AlertDialog
-import android.app.DatePickerDialog
-import android.app.TimePickerDialog
 import android.content.Intent
-import android.content.pm.PackageManager
-import android.graphics.Color
 import android.net.Uri
 import android.os.Bundle
-import android.text.Editable
-import android.text.TextWatcher
-import android.util.TypedValue
-import android.view.View
-import android.widget.*
-import androidx.appcompat.app.AppCompatActivity
-import androidx.core.app.ActivityCompat
-import com.arkadiusz.dayscounter.R
+import android.widget.TextView
 import com.arkadiusz.dayscounter.data.model.Event
-import com.arkadiusz.dayscounter.ui.internetgallery.InternetGalleryActivity
-import com.arkadiusz.dayscounter.ui.localgallery.GalleryActivity
-import com.arkadiusz.dayscounter.util.DateUtils.calculateDate
 import com.arkadiusz.dayscounter.util.DateUtils.formatDate
 import com.arkadiusz.dayscounter.util.DateUtils.formatDateAccordingToSettings
-import com.arkadiusz.dayscounter.util.DateUtils.formatTime
 import com.arkadiusz.dayscounter.util.DateUtils.generateTodayCalendar
-import com.arkadiusz.dayscounter.util.FontUtils
 import com.arkadiusz.dayscounter.util.PreferenceUtils.defaultPrefs
 import com.arkadiusz.dayscounter.util.PreferenceUtils.get
 import com.arkadiusz.dayscounter.util.PreferenceUtils.set
-import com.arkadiusz.dayscounter.util.PurchasesUtils.displayPremiumInfoDialog
-import com.arkadiusz.dayscounter.util.PurchasesUtils.isPremiumUser
-import com.arkadiusz.dayscounter.util.RemindersUtils
-import com.arkadiusz.dayscounter.util.StorageUtils.saveImage
 import com.arkadiusz.dayscounter.util.ThemeUtils
-import com.arkadiusz.dayscounter.util.ViewModelUtils.getViewModel
-import com.bumptech.glide.Glide
-import com.flask.colorpicker.ColorPickerView
-import com.flask.colorpicker.builder.ColorPickerDialogBuilder
 import com.google.android.gms.ads.AdRequest
 import com.google.android.gms.ads.interstitial.InterstitialAd
 import com.google.android.gms.ads.interstitial.InterstitialAdLoadCallback
 import com.theartofdev.edmodo.cropper.CropImage
 import kotlinx.android.synthetic.main.content_add.*
-import org.jetbrains.anko.*
-import java.io.File
 import java.util.*
 
-class AddActivity : AppCompatActivity() {
-
-    private lateinit var viewModel: AddEditViewModel
-
-    private var unformattedDate = ""
-
-    private lateinit var eventType: String
-    private var hasAlarm = false
-    private var selectedColor = -1
-    private var dimValue = 4
-
-    private val pickPhotoGallery = 1
-    private val pickPhotoInternet = 2
-    private val writeRequestCode = 1234
-
-    private var imageUri: Uri? = null
-    private var imageID = 2131230778
-    private var imageColor = 0
-
-    private var chosenYear = 0
-    private var chosenMonth = 0
-    private var chosenDay = 0
-
-    private var reminderDate = ""
-    private var chosenReminderYear = 0
-    private var chosenReminderMonth = 0
-    private var chosenReminderDay = 0
-    private var chosenReminderHour = 0
-    private var chosenReminderMinute = 0
-
-    private var wasTimePickerAlreadyDisplayed = false
+class AddActivity : BaseAddActivity() {
 
     private var interstitialAd: InterstitialAd? = null
 
     override fun onCreate(savedInstanceState: Bundle?) {
-        setTheme(ThemeUtils.getThemeFromPreferences(false, this))
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.content_add)
-        initViewModel()
-        receiveEventType()
         setCurrentDateInForm()
-        setUpSpinners()
-        setUpCheckboxes()
-        setUpEditTexts()
-        setUpSeekBar()
-        setUpFontPicker()
-        setUpOnClickListeners()
-        setUpImageChoosing()
         setUpAd()
     }
 
-    override fun onBackPressed() {
-        alert(R.string.add_activity_back_button_message) {
-            positiveButton("OK") {
-                super.onBackPressed()
-            }
-            negativeButton(R.string.add_activity_back_button_cancel) {
-                it.dismiss()
-            }
-        }.show()
-    }
-
-    private fun initViewModel() {
-        viewModel = getViewModel(this)
-    }
-
-    private fun receiveEventType() {
-        eventType = intent.getStringExtra("Event Type")!!
+    override fun handleSaveClick() {
+        showAd()
+        prepareEventBasedOnViews().also { event ->
+            viewModel.addEvent(event)
+            addReminder(event)
+        }
+        finish()
     }
 
     private fun setCurrentDateInForm() {
@@ -123,349 +45,57 @@ class AddActivity : AppCompatActivity() {
         val year = calendar.get(Calendar.YEAR)
         val month = calendar.get(Calendar.MONTH)
         val day = calendar.get(Calendar.DAY_OF_MONTH)
+
         chosenYear = year
         chosenMonth = month
         chosenDay = day
-        unformattedDate = formatDate(year, month, day)
+
+        date = formatDate(year, month, day)
+
         dateEditText.setText(
             formatDateAccordingToSettings(
-                unformattedDate,
+                date,
                 defaultPrefs(this)["dateFormat"] ?: ""
             )
         )
+
         eventCalculateText.text = generateCounterText()
     }
 
-    private fun setUpSpinners() {
-        val fontSizeAdapter = ArrayAdapter.createFromResource(
-            this,
-            R.array.add_activity_font_size,
-            R.layout.support_simple_spinner_dropdown_item
-        )
-        counterFontSizeSpinner.adapter = fontSizeAdapter
-        titleFontSizeSpinner.adapter = fontSizeAdapter
-        counterFontSizeSpinner.setSelection(6)
-        titleFontSizeSpinner.setSelection(5)
+    private fun setUpAd() {
+        val prefs = defaultPrefs(this)
+        val areAdsRemoved: Boolean = prefs["ads", false] ?: false
+        val wasShown: Boolean = prefs["wasAdShown", true] ?: true
 
-        val repetitionAdapter = ArrayAdapter.createFromResource(
-            this,
-            R.array.add_activity_repeat,
-            R.layout.support_simple_spinner_dropdown_item
-        )
-        repeatSpinner.adapter = repetitionAdapter
-        repeatSpinner.setSelection(0)
-
-        val fontTypeAdapter = FontTypeSpinnerAdapter(
-            this,
-            R.layout.support_simple_spinner_dropdown_item,
-            resources.getStringArray(R.array.font_type).toList()
-        )
-        fontTypeSpinner.adapter = fontTypeAdapter
-        fontTypeSpinner.setSelection(8)
-
-        setSpinnersListeners()
-    }
-
-    private fun setSpinnersListeners() {
-        counterFontSizeSpinner.onItemSelectedListener =
-            object : AdapterView.OnItemSelectedListener {
-                override fun onItemSelected(
-                    parent: AdapterView<*>?,
-                    view: View?,
-                    position: Int,
-                    id: Long
-                ) {
-                    if ((view as? TextView)?.text != null) {
-                        eventCalculateText.setTextSize(
-                            TypedValue.COMPLEX_UNIT_DIP,
-                            (view as? TextView)?.text.toString().toFloat()
-                        )
+        val adUnitId = "ca-app-pub-4098342918729972/3144606816"
+        if (!areAdsRemoved && !wasShown) {
+            InterstitialAd.load(
+                this,
+                adUnitId,
+                AdRequest.Builder().build(),
+                object : InterstitialAdLoadCallback() {
+                    override fun onAdLoaded(ad: InterstitialAd) {
+                        interstitialAd = ad
                     }
                 }
-
-                override fun onNothingSelected(parent: AdapterView<*>?) {
-                    //nothing
-                }
-            }
-
-        titleFontSizeSpinner.onItemSelectedListener = object : AdapterView.OnItemSelectedListener {
-            override fun onItemSelected(
-                parent: AdapterView<*>?,
-                view: View?,
-                position: Int,
-                id: Long
-            ) {
-                if ((view as? TextView)?.text != null) {
-                    eventTitle.setTextSize(
-                        TypedValue.COMPLEX_UNIT_DIP,
-                        (view as? TextView)?.text.toString().toFloat()
-                    )
-                }
-            }
-
-            override fun onNothingSelected(parent: AdapterView<*>?) {
-                //nothing
-            }
-        }
-        fontTypeSpinner.onItemSelectedListener = object : AdapterView.OnItemSelectedListener {
-            override fun onItemSelected(
-                parent: AdapterView<*>?,
-                view: View?,
-                position: Int,
-                id: Long
-            ) {
-                if ((view as? TextView)?.text != null) {
-                    val fontName = (view as? TextView)?.text.toString()
-                    val typeFace = FontUtils.getFontFor(fontName, this@AddActivity)
-                    eventTitle.typeface = typeFace
-                    eventCalculateText.typeface = typeFace
-                }
-            }
-
-            override fun onNothingSelected(parent: AdapterView<*>?) {
-                //nothing
-            }
+            )
         }
     }
 
-
-    private fun setUpCheckboxes() {
-        showDividerCheckbox.isChecked = true
-        showDividerCheckbox.setOnCheckedChangeListener { _, isChecked ->
-            if (isChecked) {
-                eventLine.visibility = View.VISIBLE
-            } else {
-                eventLine.visibility = View.GONE
-            }
-        }
-        daysCheckbox.isChecked = true
-
-        yearsCheckbox.setOnCheckedChangeListener(checkBoxListener)
-        monthsCheckbox.setOnCheckedChangeListener(checkBoxListener)
-        weeksCheckbox.setOnCheckedChangeListener(checkBoxListener)
-        daysCheckbox.setOnCheckedChangeListener(checkBoxListener)
+    private fun showAd() {
+        interstitialAd?.show(this)
+        defaultPrefs(this)["wasAdShown"] = interstitialAd != null
     }
-
-    private val checkBoxListener = CompoundButton.OnCheckedChangeListener { _, _ ->
-        if (!yearsCheckbox.isChecked && !monthsCheckbox.isChecked && !weeksCheckbox.isChecked && !daysCheckbox.isChecked) {
-            toast(getString(R.string.add_activity_toast_checkbox))
-            daysCheckbox.isChecked = true
-        }
-        eventCalculateText.text = generateCounterText()
-    }
-
-    private fun setUpEditTexts() {
-        titleEditText.addTextChangedListener(object : TextWatcher {
-            override fun afterTextChanged(s: Editable?) {
-                s?.let {
-                    eventTitle.text = it.toString()
-                }
-            }
-
-            override fun beforeTextChanged(s: CharSequence?, start: Int, count: Int, after: Int) {
-                //not used
-            }
-
-            override fun onTextChanged(s: CharSequence?, start: Int, before: Int, count: Int) {
-                //not used
-            }
-        })
-    }
-
-    private fun setUpSeekBar() {
-        pictureDimSeekBar.setOnSeekBarChangeListener(object : SeekBar.OnSeekBarChangeListener {
-            override fun onProgressChanged(seekBar: SeekBar?, progress: Int, fromUser: Boolean) {
-                eventImage.setColorFilter(Color.argb(255 / 17 * progress, 0, 0, 0))
-                dimValue = progress
-            }
-
-            override fun onStartTrackingTouch(seekBar: SeekBar?) {
-                //not used
-            }
-
-            override fun onStopTrackingTouch(seekBar: SeekBar?) {
-                //not used
-            }
-        })
-    }
-
-    private fun setUpFontPicker() {
-        colorImageView.setOnClickListener {
-            displayColorPicker()
-        }
-    }
-
-    private fun displayColorPicker() {
-        val picker = ColorPickerDialogBuilder
-            .with(this)
-            .setTitle("Choose color")
-            .wheelType(ColorPickerView.WHEEL_TYPE.CIRCLE)
-            .density(10)
-            .setPositiveButton("ok", { _, selectedColor, _ -> changeWidgetsColors(selectedColor) })
-            .setNegativeButton("cancel", { _, _ -> })
-
-        if (selectedColor != -1) picker.initialColor(selectedColor)
-
-        picker.build()
-            .show()
-    }
-
-    private fun displayColorPickerForEventBackground() {
-        val picker = ColorPickerDialogBuilder
-            .with(this)
-            .setTitle("Choose color")
-            .wheelType(ColorPickerView.WHEEL_TYPE.CIRCLE)
-            .density(10)
-            .setPositiveButton("ok", { _, selectedColor, _ -> changeEventColor(selectedColor) })
-            .setNegativeButton("cancel", { _, _ -> })
-
-        if (imageColor != -0) picker.initialColor(imageColor)
-
-        picker.build()
-            .show()
-    }
-
-    private fun changeWidgetsColors(color: Int) {
-        selectedColor = color
-        eventTitle.textColor = color
-        eventCalculateText.textColor = color
-        colorImageView.backgroundColor = color
-        eventLine.backgroundColor = color
-    }
-
-    private fun changeEventColor(color: Int) {
-        imageID = 0
-        imageUri = null
-        imageColor = color
-        eventImage.setImageDrawable(null)
-        eventImage.backgroundColor = color
-    }
-
-    private fun setUpOnClickListeners() {
-        dateEditText.setOnClickListener(showDatePicker)
-        reminderDateEditText.setOnClickListener(showReminderDatePicker)
-        clearReminderDateButton.setOnClickListener {
-            reminderDateEditText.setText("")
-            chosenReminderYear = 0
-            chosenReminderMonth = 0
-            chosenReminderDay = 0
-            chosenReminderHour = 0
-            chosenReminderMinute = 0
-            hasAlarm = false
-        }
-        addButton.setOnClickListener {
-            showAd()
-            val eventToBeAdded = prepareEventBasedOnViews()
-            viewModel.addEvent(eventToBeAdded)
-            addReminder(eventToBeAdded)
-            finish()
-        }
-    }
-
-    private val showDatePicker = View.OnClickListener {
-        val calendar = Calendar.getInstance()
-        val year = if (chosenYear == 0) calendar.get(Calendar.YEAR) else chosenYear
-        val month = if (chosenYear == 0) calendar.get(Calendar.MONTH) else chosenMonth
-        val day = if (chosenYear == 0) calendar.get(Calendar.DAY_OF_MONTH) else chosenDay
-
-        DatePickerDialog(
-            this,
-            DatePickerDialog.OnDateSetListener { _, chosenYear, chosenMonth, chosenDay ->
-                this.chosenYear = chosenYear
-                this.chosenMonth = chosenMonth
-                this.chosenDay = chosenDay
-                unformattedDate = formatDate(chosenYear, chosenMonth, chosenDay)
-                dateEditText.setText(
-                    formatDateAccordingToSettings(
-                        unformattedDate,
-                        defaultPrefs(this)["dateFormat"] ?: ""
-                    )
-                )
-                eventCalculateText.text = generateCounterText()
-            },
-            year,
-            month,
-            day
-        ).show()
-    }
-
-    private val showReminderDatePicker = View.OnClickListener {
-        val calendar = Calendar.getInstance()
-        val year = if (chosenReminderYear == 0) calendar.get(Calendar.YEAR) else chosenReminderYear
-        val month =
-            if (chosenReminderYear == 0) calendar.get(Calendar.MONTH) else chosenReminderMonth
-        val day =
-            if (chosenReminderYear == 0) calendar.get(Calendar.DAY_OF_MONTH) else chosenReminderDay
-
-        wasTimePickerAlreadyDisplayed = false
-        DatePickerDialog(
-            this,
-            DatePickerDialog.OnDateSetListener { _, chosenYear, chosenMonth, chosenDay ->
-                reminderDate = formatDate(chosenYear, chosenMonth, chosenDay)
-                chosenReminderYear = chosenYear
-                chosenReminderMonth = chosenMonth
-                chosenReminderDay = chosenDay
-                if (!wasTimePickerAlreadyDisplayed) {
-                    displayTimePickerDialog()
-                }
-            },
-            year,
-            month,
-            day
-        ).show()
-    }
-
-    private fun displayTimePickerDialog() {
-        val calendar = Calendar.getInstance()
-        val hour =
-            if (chosenReminderHour == 0) calendar.get(Calendar.HOUR_OF_DAY) else chosenReminderHour
-        val minute =
-            if (chosenReminderHour == 0) calendar.get(Calendar.MINUTE) else chosenReminderMinute
-
-        wasTimePickerAlreadyDisplayed = true
-        TimePickerDialog(
-            this,
-            TimePickerDialog.OnTimeSetListener { view, chosenHour, chosenMinute ->
-                if (view.isShown) {
-                    this.chosenReminderHour = chosenHour
-                    this.chosenReminderMinute = chosenMinute
-                    val time = formatTime(chosenHour, chosenMinute)
-                    reminderDate = formatDateAccordingToSettings(
-                        reminderDate,
-                        defaultPrefs(this)["dateFormat"] ?: ""
-                    )
-                    reminderDate += " $time"
-                    reminderDateEditText.setText(reminderDate)
-                    hasAlarm = true
-                }
-            },
-            hour,
-            minute,
-            true
-        ).show()
-    }
-
-    private fun generateCounterText(): String {
-        return calculateDate(
-            chosenYear, chosenMonth + 1, chosenDay,
-            yearsCheckbox.isChecked,
-            monthsCheckbox.isChecked,
-            weeksCheckbox.isChecked,
-            daysCheckbox.isChecked,
-            this
-        )
-    }
-
 
     private fun prepareEventBasedOnViews(): Event {
         val event = Event()
         event.name = titleEditText.text.toString()
-        event.date = unformattedDate
+        event.date = date
         event.description = descriptionEditText.text.toString()
         event.image = imageUri.toString()
         event.imageID = imageID
         event.imageColor = imageColor
-        event.type = eventType
+        event.type = intent.getStringExtra("Event Type")!!
         event.repeat = repeatSpinner.selectedItemPosition.toString()
         if (hasAlarm) {
             event.hasAlarm = true
@@ -492,118 +122,6 @@ class AddActivity : AppCompatActivity() {
         event.fontColor = selectedColor
         event.pictureDim = dimValue
         return event
-    }
-
-    private fun addReminder(eventToBeAdded: Event) {
-        if (isReminderSet()) {
-            RemindersUtils.addNewReminder(this, eventToBeAdded)
-        }
-    }
-
-    private fun isReminderSet(): Boolean = chosenReminderYear != 0 && chosenReminderDay != 0
-
-    private fun setUpImageChoosing() {
-        imageChooserButton.setOnClickListener {
-            AlertDialog.Builder(this)
-                .setTitle(getString(R.string.add_activity_dialog_title))
-                .setItems(setUpImageChooserDialogOptions()) { _, which ->
-                    when (which) {
-                        0 -> askForPermissionsAndDisplayCropActivity()
-                        1 -> startActivityForResult<GalleryActivity>(
-                            pickPhotoGallery,
-                            "activity" to "Add"
-                        )
-                        2 -> displayColorPickerForEventBackground()
-                        3 -> {
-                            if (isPremiumUser(this)) {
-                                askForPermissionsAndDisplayInternetImageActivity()
-                            } else {
-                                displayPremiumInfoDialog(this)
-                            }
-                        }
-                    }
-                }.show()
-        }
-    }
-
-    private fun setUpImageChooserDialogOptions(): Array<String> {
-        val options = mutableListOf<String>(
-            getString(R.string.add_activity_dialog_option_custom),
-            getString(R.string.add_activity_dialog_option_gallery),
-            getString(R.string.add_activity_dialog_option_color),
-            getString(R.string.add_activity_dialog_option_internet)
-        )
-        return options.toTypedArray()
-    }
-
-    override fun onNewIntent(intent: Intent?) {
-        super.onNewIntent(intent)
-
-        //intent from GalleryActivity
-        imageID = intent?.getIntExtra("imageID", 0) ?: 0
-        if (imageID != 0) {
-            imageColor = 0
-            imageUri = null
-            Glide.with(this).load(imageID).into(eventImage)
-        }
-
-        //intent from InternetGalleryActivity
-        val internetImageUri = intent?.getStringExtra("internetImageUri") ?: ""
-        if (internetImageUri.isNotEmpty()) {
-            imageUri = Uri.parse(internetImageUri)
-            imageColor = 0
-            imageID = 0
-            Glide.with(this).load(File(imageUri?.path)).into(eventImage)
-        }
-    }
-
-    private fun askForPermissionsAndDisplayCropActivity() {
-        if (ActivityCompat.checkSelfPermission(
-                this,
-                Manifest.permission.READ_EXTERNAL_STORAGE
-            ) != PackageManager.PERMISSION_GRANTED
-        ) {
-            ActivityCompat.requestPermissions(
-                this,
-                arrayOf(Manifest.permission.READ_EXTERNAL_STORAGE),
-                writeRequestCode
-            )
-        } else {
-            //permission already granted, so display crop dialog
-            CropImage.startPickImageActivity(this)
-        }
-    }
-
-    private fun askForPermissionsAndDisplayInternetImageActivity() {
-        if (ActivityCompat.checkSelfPermission(
-                this,
-                Manifest.permission.READ_EXTERNAL_STORAGE
-            ) != PackageManager.PERMISSION_GRANTED
-        ) {
-            ActivityCompat.requestPermissions(
-                this,
-                arrayOf(Manifest.permission.READ_EXTERNAL_STORAGE),
-                pickPhotoInternet
-            )
-        } else {
-            startActivityForResult<InternetGalleryActivity>(pickPhotoInternet, "activity" to "Add")
-        }
-    }
-
-
-    override fun onRequestPermissionsResult(
-        requestCode: Int,
-        permissions: Array<out String>,
-        grantResults: IntArray
-    ) {
-        super.onRequestPermissionsResult(requestCode, permissions, grantResults)
-        if (requestCode == writeRequestCode && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
-            CropImage.startPickImageActivity(this)
-        }
-
-        if (requestCode == pickPhotoInternet && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
-            startActivityForResult<InternetGalleryActivity>(pickPhotoInternet, "activity" to "Add")
-        }
     }
 
     override fun onActivityResult(requestCode: Int, resultCode: Int, imageData: Intent?) {
@@ -656,46 +174,4 @@ class AddActivity : AppCompatActivity() {
     private fun isResultComingWithImageAfterCropping(requestCode: Int): Boolean {
         return requestCode == CropImage.CROP_IMAGE_ACTIVITY_REQUEST_CODE
     }
-
-    private fun displayImageHere(data: Intent?) {
-        data?.let {
-            imageColor = 0
-            imageID = 0
-            imageUri = CropImage.getActivityResult(data).uri as Uri
-            imageUri = saveImage(this, imageUri as Uri)
-            Glide.with(this).load(File(imageUri?.path)).into(eventImage)
-        }
-    }
-
-    private fun setUpAd() {
-        val adUnitId = "ca-app-pub-4098342918729972/3144606816"
-
-        val prefs = defaultPrefs(this)
-        val areAdsRemoved: Boolean = prefs["ads", false] ?: false
-        val wasShown: Boolean = prefs["wasAdShown", true] ?: true
-
-        if (!areAdsRemoved && !wasShown) {
-            InterstitialAd.load(
-                this,
-                adUnitId,
-                AdRequest.Builder().build(),
-                object : InterstitialAdLoadCallback() {
-                    override fun onAdLoaded(ad: InterstitialAd) {
-                        interstitialAd = ad
-                    }
-                }
-            )
-        }
-    }
-
-    private fun showAd() {
-        val prefs = defaultPrefs(this)
-        if (interstitialAd != null) {
-            interstitialAd!!.show(this)
-            prefs["wasAdShown"] = true
-        } else {
-            prefs["wasAdShown"] = false
-        }
-    }
-
 }
